Refactor elastic_table to separate common functionality
